completely rewrite logger parsing
The logger should not rely on the data objects only to pluck a few
columns from the packet's message. This would also mean that the logger
would fail completely because of unrelated fields being incorrect / API
change, which is defeats the whole purpose. This picks a few fields that
should be extremely stable and expects them to be there, and nothing
else, ensuring the packet (good or bad) gets logged.